pyproject.toml
[tool.poetry]
name = "db-project-2"
version = "0.1.0"
description = ""
authors = ["Enzo Bozzani <bozzanienzo@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
psycopg2 = "^2.9.10"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
